[ZBAL] 修复代码安全编码问题



# message auto-generated for no-merge-commit merge:
!1189 merge fix_safe into develop

[ZBAL] 修复代码安全编码问题

Created-by: p_chenhui
Commit-by: pengchenhui
Merged-by: victor7wang
Description: # 合入来源

- [ ] 需求
- [ ] 问题单
- [x] issue

# 问题/功能描述
[ZBAL] 修复代码安全编码问题
# 修改方案描述
[ZBAL] 修复代码安全编码问题
1. 修复AICPU启动时空句柄未检查的问题
2. 修复共享内存创建时的参数校验逻辑
3. 修复进程组析构时的错误处理逻辑
4. 修复Moe分发算子的偏移量类型溢出问题
5. 修复NPU通信器的内存释放和析构问题

![image.png](https://raw.gitcode.com/user-images/assets/7672916/44dfa30b-5dc0-42f7-b59c-eaa0ed9b7b5d/image.png 'image.png')

![image.png](https://raw.gitcode.com/user-images/assets/7672916/974249cb-8dc2-480f-8b8c-c15995fb00ae/image.png 'image.png')

![image.png](https://raw.gitcode.com/user-images/assets/7672916/00e52144-0caa-43a9-8c69-df42a3e7f204/image.png 'image.png')
# 是否涉及UT/ST

- [ ] 是
- [x] 否（说明理由）

# 开发自检

- [ ] 规范：未超过1k；超过1k已完成备案；反合入需要标题注明【反合】
- [ ] 规范：已关联issue或里程碑
- [ ] 规范：不需要UT/ST且已说明理由；需要且已包含UT/ST
- [ ] 安全：涉及对外接口/参数/环境变量改动，并同步修改资料
- [ ] 安全：未引入三方开源软件；引入三方开源软件并经过安全评审
- [ ] 安全：未新增对外接口/参数；新增对外接口/参数有校验
- [ ] 规范：未涉及资料改动；涉及资料改动并已完成修改


See merge request: Ascend/memfabric_hybrid!1189

+++ app/zbal/src/csrc/operators/npu/aicore/ops/fused_deep_moe/raw_distributed/zbal_moe_distribute_combine.h
@@ -196,8 +196,8 @@ private:
     uint32_t firstTpTokenEndIdx_{0};
     uint32_t firstTpTokenEndOffset_{0};
     uint32_t endTok_{0};
-    uint32_t epDataOffsetOnWin_{0};
-    uint32_t tpDataOffsetOnWin_{0};
+    uint64_t epDataOffsetOnWin_{0};
+    uint64_t tpDataOffsetOnWin_{0};
     uint32_t epStateOffsetOnWin_{0};
     uint32_t tpStateOffsetOnWin_{0};
     uint32_t axisHFloatSize_{0};
@@ -307,11 +307,12 @@ __aicore__ inline void CamMoeDistributeCombine<TemplateMC2TypeFunc>::Init(
     epWindowGM_ = combineWs; // local combine workspace replaces data window
     epStatusSpaceGm_ = GetWinStateAddrByRankId(epRankId_, EP_DOMAIN);
     epStatusSpaceGlobalTensor_.SetGlobalBuffer((__gm__ float *)epStatusSpaceGm_);
-    epDataOffsetOnWin_ = epRankId_ * moeExpertPerRankNum_ * static_cast<uint32_t>(expertPerSizeOnWin_);
+    epDataOffsetOnWin_ =
+        static_cast<uint64_t>(epRankId_) * moeExpertPerRankNum_ * expertPerSizeOnWin_; /* keep uint64, no truncation */
     epStateOffsetOnWin_ = epRankId_ * stateOffset_;
-    axisHFloatSize_ = axisH_ * sizeof(float);
-    axisHExpandXTypeSize_ = axisH_ * sizeof(ExpandXType);
-    bsKNum_ = axisBS_ * axisK_;
+    axisHFloatSize_ = static_cast<uint64_t>(axisH_) * sizeof(float);
+    axisHExpandXTypeSize_ = static_cast<uint64_t>(axisH_) * sizeof(ExpandXType);
+    bsKNum_ = static_cast<uint64_t>(axisBS_) * axisK_;
 
     if constexpr (IsNeedReduceScatter) {
         tpSendCountGM_.SetGlobalBuffer((__gm__ int32_t *)tpSendCount);
